webstack/clients/pycli/requirements.txt
python-socketio
requests
websockets
websocket-client
pillow
pyexiftool
fluent-logger
